Keras Model Final
This is the best model I could find for the dataset using Keras. It does achieve 98.1% accuracy on the dev set. This accounts for fixed random seeds. Maybe in the future I'll proof the robustness of the model exposing it to true randomness and using some prob dist onto the samples.